Allow array of strings as description (#23)
Semantic UI message supports array of strings as `content` prop and generates a simple list from the items. I use it quite a lot and would be nice to be able to pass it to the toast. 

Right now, I'm working around this issue by passing a `Message.List` component.